pyproject.toml
[project]
name = "wordplay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PySide6>=6.7",
    "llama-cpp-python>=0.3",
]
